COC-91-refactor(#73): StudyList 하위 폴더 명 수정
