+++ src/pages/Study/StudyList/Body/style.ts
