+++ src/pages/Study/StudyList/Footer/index.tsx
