+++ src/pages/Study/StudyList/Footer/style.ts
