+++ src/pages/Study/StudyList/Header/style.ts
