+++ src/pages/Study/StudyList/index.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import Header from './header';
-import Body from './body';
-import Footer from './footer';
+import Header from './Header';
+import Body from './Body';
+import Footer from './Footer';
 import * as S from './style';
 
 export default function StudyList() {
